fetch_image/requirements.txt
requests
boto3
numpy
pillow
face_recognition